refactor: extract section for user to change name from dashboard to own component
