+++ src/components/Dashboard/Dashboard.tsx
@@ -1,16 +1,15 @@
 // @ts-check
 import * as React from 'react';
-const {useEffect, useState} = React;
+const {useEffect} = React;
 import { signOut } from "next-auth/client";
-import { useRouter } from 'next/router'
-import Link from 'next/link';
 import styles from "../../styles/Home.module.css";
-import {logoutUser, fetchUserDisplayName, updateUserDisplayName} from '../../redux/reducers/userSlice';
+import {logoutUser, fetchUserDisplayName} from '../../redux/reducers/userSlice';
 import { useAppSelector, useAppDispatch } from '../../../lib/hooks';
 import { DashboardProps } from '../../../pages/dashboard';
 import PlayGameSection from './PlayGameSection'
 import CreateGameSection from './CreateGameSection'
 import DraftGames from './DraftGames'
+import ChangeDisplayNameSection from './ChangeDisplayNameSection'
 
 const {
   buttonContainer,
@@ -26,9 +25,6 @@ const {
 const Dashboard: React.FC <DashboardProps> = (props) => {
   const dispatch = useAppDispatch()
   const {draftGames, session} = props
-  const [userIsChangingDisplayName, setUserIsChangingDisplayName] =
-    useState(false)
-  const [displayName, setDisplayName] = useState('')
   const userDisplayName = useAppSelector((state) => state.user.userDisplayName)
   const userId = useAppSelector((state) => state.user.userId)
   useEffect(() => {
@@ -36,11 +32,6 @@ const Dashboard: React.FC <DashboardProps> = (props) => {
       dispatch(fetchUserDisplayName(session));
     }
   }, []);
-  const submitNewDisplayName = () => {
-    dispatch(updateUserDisplayName({displayName, userId}));
-    setDisplayName(null);
-    setUserIsChangingDisplayName(false);
-  };
   const initiateSignOut = () => {
     signOut();
     dispatch(logoutUser());
@@ -58,28 +49,8 @@ const Dashboard: React.FC <DashboardProps> = (props) => {
               {userDisplayName || session.user.name || session.user.email}
             </h4>
           </div>
-          <div>
-            <button
-              type="button"
-              onClick={() => setUserIsChangingDisplayName(true)}
-            >
-              Change display name
-            </button>
-          </div>
-          {userIsChangingDisplayName ? (
-            <div>
-              <label htmlFor="newDisplayName">New Display Name: </label>
-              <input
-                type="text"
-                name="displayName"
-                value={displayName}
-                onChange={(e) => setDisplayName(e.target.value)}
-              ></input>
-              <button type="button" onClick={submitNewDisplayName}>
-                Submit new display name
-              </button>
-            </div>
-          ) : null}
+          <ChangeDisplayNameSection />
+
           <div className={signOutButtonContainer}>
             <button type="button" onClick={initiateSignOut}>
               Sign out
